Dedup feeders and fountains
The majority of the various feeder and fountain classes were copied from
one another. This is an effort to make them easier to maintain by
placing the code in one place while maintaining the same interface as
before.

As with PR #11 I still don't have a testing Home Assistant setup so I
need to get that running first to verify that this actually works.