Remove sensors not available in Polar units
While the API does return values for these sensors, the device does not actually support these:

- food_low - Polar units only have a plate and don't measure how much food is in them
- whether_in_sleep_mode - The devices do not have any sleep mode

Polar units actually *do* support temperatures (visible when spying on the MQTT communication) but the API does not report it. It only supports a "snowflake" symbol in the app, but I'm not even sure what it means.