Polar: Enable new sensors and clean up existing properties
Some properties were not used/duplicate:

 - `available` was not used and was wrong - the property device does not exist
 - `device_sn` is a duplicate of `serial` from the Device class
 - `mac_address` is a duplicate of `mac` from Device. Also wasn't used
 - `temperature` was already removed as a sensor because the API does not report the temperature

Most properties were updated to return `None` if they have no value. This allows HomeAssistant to signal that the value is not available instead of hardcoding something.

The feeding times were updated to return proper datetimes and the sensors now show as timestamps. This unifies the date and time fields into one and allow users to have their own locale for the fields. It also knows about the proper timezone now. This could also allow users to properly use the date/times in automations or scripts.

Add debug logging